Refactor grimoirelab package into smaller files
